exch-publisher/requirements.txt
google-cloud-pubsub
lxml